Revise import statements in accordance with `baseUrl`
