+++ frontend/src/components/Account/Password.tsx
@@ -1,14 +1,14 @@
 import { useState } from 'react';
+
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
 import { Box, Grid, TextField } from '@material-ui/core';
-import { useAppDispatch } from '../../utils/hooks/useAppDipatch';
+
 import { updatePassword } from 'store/thunks';
-import LabeledCheckbox from '../../templates/LabeledCheckbox';
-import AlertMessage from '../../templates/AlertMessge';
-import SubmitButton from '../../templates/SubmitButton';
-import { isGuest } from '../../utils/auth';
+import { useAppDispatch } from 'utils/hooks';
+import { isGuest } from 'utils/auth';
+import { LabeledCheckbox, AlertMessage, SubmitButton } from 'templates';
 
 // Input items
 type FormData = {
