+++ frontend/src/components/Account/index.ts
@@ -1,0 +1,4 @@
+export { default as DeleteAccountDialog } from './DeleteAccountDialog';
+export { default as Password } from './Password';
+export { default as UserProfile } from './UserProfile';
+export { default as UserStatus } from './UserStatus';
