+++ frontend/src/components/Header/AccountMenuList.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { useHistory } from 'react-router-dom';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { List, ListItem, ListItemIcon, ListItemText } from '@material-ui/core';
@@ -6,8 +7,9 @@ import {
   AccountCircle as AccountCircleIcon,
   ExitToApp as ExitToAppIcon,
 } from '@material-ui/icons';
-import { useAppDispatch, useAppSelector } from 'utils/hooks';
+
 import { signOutFromAPI } from 'store/thunks';
+import { useAppDispatch, useAppSelector } from 'utils/hooks';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
