+++ frontend/src/components/Home/Hero.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
+
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { Container, Grid, Typography, Box } from '@material-ui/core';
-import hero from '../../images/hero.svg';
+
+import hero from 'images/hero.svg';
 import HeroAction from './HeroAction';
 
 const useStyles = makeStyles((theme: Theme) =>
