+++ frontend/src/components/Home/HeroAction.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { Link as RouterLink } from 'react-router-dom';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { Grid, Button } from '@material-ui/core';
@@ -6,6 +7,7 @@ import {
   PersonAdd as PersonAddIcon,
   LockOpen as LockOpenIcon,
 } from '@material-ui/icons';
+
 import GuestAction from './GuestAction';
 
 const useStyles = makeStyles((theme: Theme) =>
