+++ frontend/src/components/Home/index.ts
@@ -1,0 +1,4 @@
+export { default as Hero } from './Hero';
+export { default as HeroAction } from './HeroAction';
+export { default as GuestAction } from './GuestAction';
+export { default as Features } from './Features';
