+++ frontend/src/index.tsx
@@ -1,14 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+
 import { Provider } from 'react-redux';
 import { HelmetProvider } from 'react-helmet-async';
 import { ThemeProvider } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 
-import App from './App';
 import reportWebVitals from './reportWebVitals';
 import store from './store';
 import theme from './theme';
+import App from './App';
 
 if (process.env.NODE_ENV === 'development') {
   const { worker } = require('./mocks/browser');
