+++ frontend/src/layouts/MarkdownWithToc.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
+
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { Container, Grid, Link } from '@material-ui/core';
-import Markdown from '../templates/Markdown';
+
+import Markdown from 'templates/Markdown';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
