+++ frontend/src/layouts/Sidebar.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import {
   Typography,
