+++ frontend/src/layouts/index.ts
@@ -1,0 +1,6 @@
+export { default as Header } from './Header';
+export { default as Footer } from './Footer';
+export { default as Loading } from './Loading';
+export { default as FlashNotification } from './FlashNotification';
+export { default as FormLayout } from './FormLayout';
+export { default as MarkdownWithToc } from './MarkdownWithToc';
