+++ frontend/src/pages/auth/EmailVerification.tsx
@@ -1,15 +1,16 @@
 import React, { useEffect } from 'react';
+
 import { useHistory } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { Container, Card, Box, Typography, Button } from '@material-ui/core';
-import Header from '../../layouts/Header';
-import Footer from '../../layouts/Footer';
-import { APP_NAME } from '../../config/app';
-import { useAppDispatch, useAppSelector } from 'utils/hooks';
+
+import { APP_NAME } from 'config/app';
 import { deleteSentEmailState } from 'store/slices/authSlice';
 import { sendEmailVerificationLink } from 'store/thunks';
-import { isSentEmail, isVerified } from '../../utils/auth';
+import { useAppDispatch, useAppSelector } from 'utils/hooks';
+import { isSentEmail, isVerified } from 'utils/auth';
+import { Header, Footer } from 'layouts';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
