+++ frontend/src/pages/auth/ResetPassword.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+
 import { Helmet } from 'react-helmet-async';
 import { useHistory } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
@@ -6,13 +7,12 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { TextField, Button, Divider, Grid, Box } from '@material-ui/core';
-import { APP_NAME } from '../../config/app';
-import { useQuery } from 'utils/hooks';
-import { useAppDispatch } from 'utils/hooks';
+
+import { APP_NAME } from 'config/app';
 import { signInWithEmail, resetPassword } from 'store/thunks';
-import FormLayout from '../../layouts/FormLayout';
-import LabeledCheckbox from '../../templates/LabeledCheckbox';
-import SubmitButton from 'templates/SubmitButton';
+import { useAppDispatch, useQuery } from 'utils/hooks';
+import { FormLayout } from 'layouts';
+import { LabeledCheckbox, SubmitButton } from 'templates';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
