+++ frontend/src/templates/Markdown.tsx
@@ -1,4 +1,5 @@
 import { ReactNode } from 'react';
+
 import MarkdownToJsx, { MarkdownToJSX } from 'markdown-to-jsx';
 import { Typography } from '@material-ui/core';
