+++ frontend/src/templates/PopoverControl.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { Popover, Box } from '@material-ui/core';
