+++ frontend/src/templates/index.ts
@@ -1,0 +1,4 @@
+export { default as AlertMessage } from './AlertMessge';
+export { default as LabeledCheckbox } from './LabeledCheckbox';
+export { default as PopoverControl } from './PopoverControl';
+export { default as SubmitButton } from './SubmitButton';
