+++ frontend/src/theme/index.ts
@@ -1,4 +1,5 @@
 import { createMuiTheme } from '@material-ui/core/styles';
+
 import overrides from './overrides';
 import typography from './typography';
