+++ frontend/src/utils/generator/index.ts
@@ -1,0 +1,2 @@
+export * from './generateRandomString';
+export * from './makeEmail';
