+++ frontend/src/utils/generator/makeEmail.ts
@@ -1,0 +1,8 @@
+import { generateRandomString } from '.';
+
+export const makeEmail = () => {
+  const username = generateRandomString() + generateRandomString();
+  const domain = 'example.com';
+  const email = username + '@' + domain;
+  return email;
+};
